+++ Cinemastic/src/Core/Cinemastic.Domain/Entities/ContentGenre.cs
@@ -8,8 +8,8 @@ namespace Cinemastic.Domain.Entities
 {
     public class ContentGenre
     {
-        public int ContentId { get; set; }
-        public int GenreId { get; set; }
+        public long ContentId { get; set; }
+        public long GenreId { get; set; }
         public Content Content { get; set; }
         public Genre Genre { get; set; }
     }
