+++ Cinemastic/src/Core/Cinemastic.Domain/Entities/ContentTag.cs
@@ -8,8 +8,8 @@ namespace Cinemastic.Domain.Entities
 {
     public class ContentTag
     {
-        public int ContentId { get; set; }
-        public int TagId { get; set; }
+        public long ContentId { get; set; }
+        public long TagId { get; set; }
         public Content Content { get; set; }
         public Tag Tag { get; set; }
     }
